backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
motor==3.3.1
pydantic==2.4.2
pydantic[email]==2.4.2
python-dotenv==1.0.0
websockets==11.0.3
docker==6.1.3
pymongo==4.5.0
python-multipart==0.0.6
pytest==7.4.2
httpx==0.25.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0